Dash Code & Resources/authentication/requirements.txt
pandas
plotly
numpy
dash_auth
python-decouple
gunicorn